chore: Refactor jsonrpc server into its own package
